backend/ms-venta/requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
pytest==8.2.0
pytest-django==4.8.0
coverage==7.5.1

django-cors-headers==4.7.0